### 2021-05-05 #112 @whedon
[PRE REVIEW]: EPIDEMIC: Epidemiology Educational Code
**Submitting author:** @BrunaPavlack (<a href="http://orcid.org/0000-0002-6807-0916">Bruna Silveira Pavlack</a>)
**Repository:** <a href="https://github.com/americocunhajr/EPIDEMIC" target ="_blank">https://github.com/americocunhajr/EPIDEMIC</a>
**Version:** 1.0.0

@BrunaPavlack if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
