### 2021-05-24 #112 @labarba
@BrunaPavlack 👋  — Your paper is not compiling because the metadata is nonconforming to the specs. Please check the [example paper](https://openjournals.readthedocs.io/en/jose/submitting.html#example-paper-and-bibliography) and fix.
