### 2021-06-02 #112 @BrunaPavlack
> @BrunaPavlack 👋 — Your paper is not compiling because the metadata is nonconforming to the specs. Please check the [example paper](https://openjournals.readthedocs.io/en/jose/submitting.html#example-paper-and-bibliography) and fix.

Hello! I would like help identifying what the problem is with compiling the paper. I used the site https://whedon.theoj.org/  to view and compiled it just right. Thank you very much in advance!
