### 2021-06-04 #112 @arfon
> Author (Amanda Cunha Guyt) is missing affiliation

Please add an affiliation for Amanda Cunha Guyt as per the error above.
