### 2021-06-14 853e731
Update rhs_SEIAHRD.m
