### 2021-06-14 d40bd97
Update main_SEIAHRD_covid.m
