### 2021-06-14 8c5b222
Update check_error_SEIAHRD.m
