+++ modeling/SEIAHRD/check_error_SEIAHRD.m
@@ -44,17 +44,17 @@ disp(' -----------------------------------------------')
 disp(' ++++++++++++ rhs_SEIAHRD checking +++++++++++++')
 % -----------------------------------------------------------
 
-% Case 1: length(param) < 9 
+% Case 1: length(param) < 8
 % -----------------------------------------------------------  
 t = 1;                                      % time
 y = [999 0 1 0 0 0 0 1];                    % state vector
-param = [1000 0.5 0.5 0.1 0.1 0.5 0.2 0.3]; % param. vector
+param = [1000 0.5 0.1 0.1 0.5 0.2 0.3]; % param. vector
 
 try  
     rhs_SEIAHRD(t,y,param)
 catch ME
     disp(' ')
-    disp('Case 1: length(param) < 9')
+    disp('Case 1: length(param) < 8')
     disp(['t: ',num2str(t)])
     disp(['y: ',num2str(y)])
     disp(['param: ',num2str(param)])
@@ -66,15 +66,15 @@ catch ME
 end
 % ----------------------------------------------------------- 
 
-% Case 2: length(param) > 9
+% Case 2: length(param) > 8
 % -----------------------------------------------------------  
-param = [1000 0.5 0.5 0.1 0.1 0.5 0.2 0.3 0.2 0.2]; % parameters vector
+param = [1000 0.5 0.1 0.1 0.5 0.2 0.3 0.2 0.2]; % parameters vector
 
 try  
     rhs_SEIAHRD(t,y,param)
 catch ME
     disp(' ')
-    disp('Case 2: length(param) > 9')
+    disp('Case 2: length(param) > 8')
     disp(['t: ',num2str(t)])
     disp(['y: ',num2str(y)])
     disp(['param: ',num2str(param)])
@@ -88,7 +88,7 @@ end
 
 % Case 3: mod(param(1),1) ~= 0
 % -----------------------------------------------------------  
-param = [1000.25 0.5 0.5 0.5 0.1 0.3 0.2 0.2 0.2]; % parameters vector
+param = [1000.25 0.5 0.5 0.1 0.3 0.2 0.2 0.2]; % parameters vector
 
 try  
     rhs_SEIAHRD(t,y,param)
@@ -108,7 +108,7 @@ end
 
 % Case 4: param(1) < 0
 % -----------------------------------------------------------  
-param = [-1000 0.5 0.5 0.5 0.1 0.3 0.2 0.2 0.2]; % parameters vector
+param = [-1000 0.5 0.5 0.1 0.3 0.2 0.2 0.2]; % parameters vector
 
 try  
     rhs_SEIAHRD(t,y,param)
@@ -128,7 +128,7 @@ end
 
 % Case 5: param(2) < 0
 % -----------------------------------------------------------  
-param = [1000 -0.5 0.5 0.5 0.1 0.3 0.2 0.2 0.2]; % parameters vector
+param = [1000 -0.5 0.5 0.1 0.3 0.2 0.2 0.2]; % parameters vector
 
 try  
     rhs_SEIAHRD(t,y,param)
@@ -148,7 +148,7 @@ end
 
 % Case 6: param(3) < 0
 % -----------------------------------------------------------  
-param = [1000 0.5 -0.5 0.5 0.1 0.3 0.2 0.2 0.2]; % parameters vector
+param = [1000 0.5 -0.5 0.1 0.3 0.2 0.2 0.2]; % parameters vector
 
 try  
     rhs_SEIAHRD(t,y,param)
@@ -168,7 +168,7 @@ end
 
 % Case 7: param(4) < 0
 % -----------------------------------------------------------  
[…]
