### 2021-06-14 c7c8dde
Update check_error_SEIAHRD.m
