### 2021-06-14 450e655
Update rhs_SEIAHRD.m
