### 2021-06-28 #112 @magsol
@BrunaPavlack The reviewers you suggested are excellent open science researchers, but none of them seem to have experience in computational epidemiology. Are you sure you'd like these folks to act as reviewers?
