### 2021-06-28 #112 @BrunaPavlack
> @BrunaPavlack The reviewers you suggested are excellent open science researchers, but none of them seem to have experience in computational epidemiology. Are you sure you'd like these folks to act as reviewers?

Helo! I could not identify any with experience in the epidemiological area, feel free to suggest new reviewers.
