### 2021-07-13 #112 @labarba
hi @BrunaPavlack 👋
I just read your paper and I have the following comment: I miss a passage that explains to readers how they might potentially use the software in their teaching and learning context. If you have used it within a course, it helps to describe briefly how it was used. Mostly, though, what we want is to make it easier for another instructor or a self-learner to visualize how they might reuse your artifacts. Have a look at the ["What should my paper contain?"](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain) of our documentation.
