### 2021-07-16 #112 @BrunaPavlack
> hi @BrunaPavlack 👋
> I just read your paper and I have the following comment: I miss a passage that explains to readers how they might potentially use the software in their teaching and learning context. If you have used it within a course, it helps to describe briefly how it was used. Mostly, though, what we want is to make it easier for another instructor or a self-learner to visualize how they might reuse your artifacts. Have a look at the ["What should my paper contain?"](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain) of our documentation.

Hi! Thank you so much for the comment! We have already added in the paper the potential use of EPIDEMIC for teaching and learning. We've also added an example of using EPIDEMIC in a course on numerical and computational
methods. We are available for any questions!
