### 2021-07-16 #112 @BrunaPavlack
Hi @labarba  ! Thank you so much for the comment! We have already added in the paper the potential use of EPIDEMIC for teaching and learning. We've also added an example of using EPIDEMIC in a course on numerical and computational
methods. We are available for any questions!
