+++ modeling/SEIAHRD/main_SEIRAHD_covid.m
@@ -1,0 +1,284 @@
+% -----------------------------------------------------------
+% EPIDEMIC - Epidemiology Educational Code
+% www.EpidemicCode.org
+% -----------------------------------------------------------
+% Modeling: main_SEIR(+AHD).m
+%
+% This is the main file for the SEIR(+AHD) epidemic model, which
+% divides a population in 7 compartments:
+%
+%   S = susceptibles
+%   E = exposed
+%   I = symptomatic infectious
+%   A = asymptomatic infectious
+%   H = hospitalized
+%   R = recovered
+%   D = deceased
+%
+% Infection spreads via direct contact between
+% a susceptible and infectious individual.
+% Delay is modeled as an exposed group: there is an
+% latent period until an infected becomes able to transmit.
+% Among the infectious, most individuals are asymptomatic; only
+% a small fraction display symptoms after incubation.
+% Disease-related deaths are considered when infectious.
+% A control procedure is considered: a fraction of the infectious,
+% is hospitalized, thus reducing their infectivity and fatality chance.
+%
+% This model has 9 parameters:
+%
+%   N0       = initial population size            (number of individuals)
+%   beta     = transmission rate                  (days^-1)
+%   alpha    = latent rate                        (days^-1)
+%   fE       = symptomatic fraction               (dimensionless)
+%   gamma    = recovery rate                      (days^-1)
+%   rho      = hospitalization rate               (days^-1)
+%   delta    = death rate                         (days^-1)
+%   kappaH   = hospitalization mortality-factor   (dimensionless)
+%
+% This codes uses rhs_SEIR(+AHD).m to define the ODE system
+% and outputs the plots, R_nought value and R_control value. 
+% Calculations are made on a day time scale.
+%
+% The parameters used in this example were taken from this work.
+% W. Lyra, J. Nascimento, J. Belkhiria, L. de Almeida,
+% P. P. Chrispim and I. Andrade.COVID-19 pandemics modelingwith
+% SEIR(+CAQH), social distancing, and age stratification. The effect
+% of vertical confinement and release in Brazil.,medRxiv, 2020.
+%
+% Inputs:
+%   param: parameters vector      - double array (9x1)
+%   IC: initial conditions vector - double array (8X1)
+%   tspan: time interval          - double array (?x1)
+%   rhs_SEIR(+AHD): SEIR(+AHD) equations file - .m function file
+%
+% Outputs:
+%   R_nought: basic reproduction number   - double
+%   figure 1: model state in time         - inplace figure
+%   figure 2: number of new cases in time - inplace figure
+% -----------------------------------------------------------
+% programmers: Eber Dantas
+%              Americo Cunha
+%
+% number of lines: 119
+% last update: Jan 26, 2021
+% -----------------------------------------------------------
+
+clc
+clear
+close all
+
+
+% parameters and initial conditions [USER INPUT]
+% -----------------------------------------------------------  
+
+% initial population size (number of individuals)
+N0 = 1000;
+        
+% transmission rate (days^-1)
+beta = 0.25;
+
+% hospitalization infectivity-factor (dimensionless)
+%
+% -- Models contact diminishment. 
+
+% latent period (days)
+Talpha = 5;
+
+% latent rate (days^-1)
+alpha = 1/Talpha;
+
+% symptomatic fraction (dimensionless)
+%
+% -- Models fraction of infectious that display symptoms. 
+% -- Values: 0<fE<1.
+fE = 0.6;
+
+% recovery period (days)
+Tgamma = 14;
+
+% recovery rate (days^-1)
+gamma = 1/Tgamma;
+
+% hospitalization rate (days^-1)
+rho = 0.1;
+
[…]

[…]
