### 2021-09-30 71ebaae
Delete check_error_SEIAHRD.m
