### 2021-09-30 5c5bf50
Delete main_SEIAHRD.m
