### 2021-09-30 7467f8b
Delete main_SEIAHRD_covid.m
