### 2021-09-30 c6fde83
Delete rhs_SEIAHRD.m
