### 2021-10-05 3ba0790
Update epidemic_trends.m
