+++ trends/epidemic_trends.m
@@ -530,7 +530,7 @@ ylabel ({'Total cases ','(per million inhabitants)'},...
 legend ('location','northeastoutside');
 y_init = 10;
 max_y  = 10000000;
-axis([0 day_axis y_init max_y]);
+axis([0 day_axis+50 y_init max_y]);
 
 figure(9)
 set(gca,'FontSize',font_default)
@@ -562,7 +562,7 @@ ylabel ({'New cases per week','(per million inhabitants)'},...
 legend ('location','northeastoutside');
 y_init = 0.1;
 max_y  = 1000000;
-axis([0 day_axis/7 y_init max_y]);
+axis([0 day_axis/7+7 y_init max_y]);
 
 % Saving the charts
 % -----------------------------------------------------------------------
