### 2021-10-07 4850454
Update epidemic_trends.m
