### 2021-10-16 #112 @magsol
Hi @BrunaPavlack! I'm so sorry this has been taking awhile, but I've been having a lot of trouble finding reviewers. I've been emailing potential reviewers offline, and so far have contacted 4 and had 4 subsequent declines :( I'm still following up on some additional reviewer leads, and if you have any suggestions for people who could potentially review your submission I'd be happy to hear them!

Otherwise I will certainly update you again next week with whatever progress I have made in finding reviewers.
