### 2021-11-04 #112 @labarba
Maybe `@lwjohnst86` could volunteer a review here, as the author of the recently published `r-cubed` paper: https://jose.theoj.org/papers/10.21105/jose.00122
