### 2021-11-11 #149 @whedon
[REVIEW]: EPIDEMIC: Epidemiology Educational Code
**Submitting author:** @BrunaPavlack (<a href="http://orcid.org/0000-0002-6807-0916">Bruna Silveira Pavlack</a>)
**Repository:** <a href="https://github.com/americocunhajr/EPIDEMIC" target ="_blank">https://github.com/americocunhajr/EPIDEMIC</a>
**Version:** v2.0.0

- [x] #150

- [x] **Installation:** Does installation proceed as outlined in the documentation? (and documentation is sufficient?)  **unable to assess**
- [x] **Functionality:** Have the functional claims of the software been confirmed? **unable to assess**
- [x] **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.) **unable to assess**

- [x] **Installation instructions:** Is there a clearly stated list of dependencies? (Ideally these should be handled with an automated package management solution.) **instructions could not be found, issue submitted**

- [x] **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)? **Chatterjee2020 link non-functional, otherwise yes**
