### 2021-11-12 d28f2d3
Update paper.md
minor typo
