### 2021-11-14 #149 @BrunaPavlack
> I can find no installation instructions in the Readme, the paper or the tutorial

Hi! First of all, thank you so much for agreeing to review our code @3mmaRand! The code runs in MATLAB or GNU Octave software. Put the MATLAB working directory where the code is and run via terminal. Let's add in the materials how to install and use. Thank you!
