### 2021-11-18 #149 @BrunaPavlack
Hi! We added a part about installation and use in the tutorial and in the read me. Thanks for the comment!
