### 2021-12-18 #149 @3mmaRand
hello @magsol @BrunaPavlack The paper describes the software as aimed at " ..... students who are interested in computational epidemiology. The programs and tutorials are designed to offer good introductory material." I have lots of biology students who would be interested in this! But they wouldn't be able to use it because they would need more instruction for using matlab/octave. So maybe describe the software as aimed at matlab users intersted in computational epidemiology or add a bit more instruction>
