### 2021-12-20 #149 @BrunaPavlack
> hello @magsol @BrunaPavlack The paper describes the software as aimed at " ..... students who are interested in computational epidemiology. The programs and tutorials are designed to offer good introductory material." I have lots of biology students who would be interested in this! But they wouldn't be able to use it because they would need more instruction for using matlab/octave. So maybe describe the software as aimed at matlab users intersted in computational epidemiology or add a bit more instruction>

Hello @3mmaRand ! Thank you very much for your observation! We have already updated the paper the fact that it is interesting that the user has previous knowledge of Octave or MATLAB. On the other hand, as the organization of the EPIDEMIC code is done in a very didactic way, those students who have first access to these programming languages ​​may also be able to reproduce the simulations. Little by little, we are always updating the tutorial and adding more information about the use of GNU Octave and MATLAB software.
