### 2021-12-21 #149 @dalonsoa
I have a few comments, mostly focussed on the code and the documentation. I have not checked the paper itself, yet:

## Version
There are no tags or releases in the GitHub repo matching the version indicated up in this thread (1.0.0). They should be created.

Although not required by JOSE, I would recomend getting a DOI for the tool via Zenodo or Figshare, for example, to help citing and reproducibility.

## Installation instructions
The documentation file says in page 9:

> After downloading, just run the script. The working directory of compatible software is placed in the place where the EPIDEMIC code is and it is executed via the terminal.

This is not helpful. What script should be run? From where? What "directory of compatible software" means? The next page shows a screenshot, but still it does not provide proper instructions on how to use the tool, specially for those unfamiliar with how Octave or Matlab operate.

There should also be an statement about what version(s) of Matlab or Octave are required to run the tool or, at the very least, in which versions it has been run. If extra packages or toolkits are required, it also needs to be mentioned.

## Documentation content and CONTRIBUTING guidelines

Keeping aside improving the installation/usage instructions, I miss documenting the API and how the code itself is structured, so potential contributiors know where to find the different components and improve them or add new features. In this respect, there is no CONTRIBUTING guidelines, neither. There is a message at the very end of the README indicating to send an email, but to me that is inssuficient for an open source software.

## Testing

There are no test, manual or automated, to run and prove that things work as expected. [Octave has an easy way of including tests](https://wiki.octave.org/Tests), so I strongly recomend the authors make use of it throughout the code and either setup an automated testing workflow using GitHub Actions or provide instructions on how to run the tests.

## Documentation format

Finally, having the documentation only as a monolythinc PDF file is pretty hard to use. It is not easy to navigate to the section of interest, even though it has a ToC that you can use to jump to one place to another. Changing that to a more accesible format (online, in GitHub pages which is already setup, or Read the Docs, for example) will be really useful - although I understand it is a lot of work. Not necessary now, but really something to consider in the future.
