+++ paper/paper.bib
@@ -20,8 +20,8 @@
     site = {Towards Data Science},
     title = {Transparency, Reproducibility, and Validity of {COVID-19} Projection Models},
     year = {2020},
-    url = {encurtador.com.br/dgno9},
-    urldate = {2020-06-30}
+    url = {encurtador.com.br/kmSW3},
+    urldate = {2020-05-04}
 }
 
 @online{covid19-rj,
