### 2022-01-17 #149 @BrunaPavlack
> Hey folks! Any progress here on implementing the reviewer's feedback?

Hello! sorry for the delay in getting back. We are finalizing implementations of feedback received by reviewers. We have already updated some suggestions in the tutorial. We've added more installation instructions and also added in readme more information about contributions. We need to finish the question of updating the tests and also the version to put in ZENODO. About the format of the tutorial, we are already discussing to update according to the suggestions, however we will do more in the future. Any new information will be posted here. Thanks!
