### 2022-01-20 8bf92cf
Update main_SIR.m
