### 2022-01-20 1372b22
Update main_SIR_Brauer.m
