+++ modeling/SEIR/main_SEIR.m
@@ -198,6 +198,8 @@ hold off
     xlim([t0 t1]);
     ylim([0 N]);
 
+    saveas(figure(1),'fig_modeling_SEIR_example1_compart.png')
+
 
 % plot NewCases (per day) of SEIR model
 figure(2)
@@ -214,4 +216,6 @@ fig2 = scatter([1:length(NewCases)]+1,NewCases);
 
     % axis limits
     xlim([t0 length(NewCases)]+1);
+
+    saveas(figure(2),'fig_modeling_SEIR_example1_newcases.png')
 % -----------------------------------------------------------
