+++ modeling/SEIR/main_SEIR_Dantas.m
@@ -205,6 +205,8 @@ hold off
     xlim([t0 t1]);
     ylim([0 N]);
 
+    saveas(figure(1),'fig_modeling_SEIR_example2_compart.png')
+
 
 % plot NewCases (per day) of SEIR model
 figure(2)
@@ -221,4 +223,6 @@ fig2 = scatter([1:length(NewCases)]+1,NewCases);
 
     % axis limits
     xlim([t0 length(NewCases)]+1);
+
+    saveas(figure(2),'fig_modeling_SEIR_example2_newcases.png')
 % -----------------------------------------------------------
