### 2022-01-20 78a8f5d
Update main_SIR_Brauer.m
