+++ modeling/SIR/main_SIR_Brauer.m
@@ -185,7 +185,7 @@ hold off
     xlim([t0 t1]);
     ylim([0 N]);
 
-    saveas(figure(1),'fig_modeling_SIR_example1_compart.png')
+    saveas(figure(1),'fig_modeling_SIR_example2_compart.png')
 
 
 % plot NewCases (per day) of SIR model
@@ -204,7 +204,7 @@ fig2 = scatter([1:length(NewCases)]+1,NewCases);
     % axis limits
     xlim([t0 length(NewCases)]+1);
 
-    saveas(figure(2),'fig_modeling_SIR_example1_newcases.png')
+    saveas(figure(2),'fig_modeling_SIR_example2_newcases.png')
 % -----------------------------------------------------------
 
 
