+++ modeling/SEIRD/main_SEIRD.m
@@ -209,6 +209,8 @@ hold off
     xlim([t0 t1]);
     ylim([0 N0]);
 
+    saveas(figure(1),'fig_modeling_SEIRD_example1_compart.png')
+
 
 % plot NewCases (per day) of SEIRD model
 figure(2)
@@ -225,4 +227,6 @@ fig2 = scatter([1:length(NewCases)]+1,NewCases);
 
     % axis limits
     xlim([t0 length(NewCases)]+1);
+
+    saveas(figure(2),'fig_modeling_SEIRD_example1_newcases.png')
 % -----------------------------------------------------------
