### 2022-01-20 98c333f
Update main_SEIRD_ebola.m
