+++ modeling/SEIRD/main_SEIRD_ebola.m
@@ -216,6 +216,8 @@ hold off
     xlim([t0 t1]);
     ylim([0 N0]);
 
+    saveas(figure(1),'fig_modeling_SEIRD_example2_compart.png')
+
 
 % plot NewCases (per day) of SEIRD model
 figure(2)
@@ -232,4 +234,6 @@ fig2 = scatter([1:length(NewCases)]+1,NewCases);
 
     % axis limits
     xlim([t0 length(NewCases)]+1);
+
+    saveas(figure(2),'fig_modeling_SEIRD_example2_newcases.png')
 % -----------------------------------------------------------
