### 2022-01-20 5e731a9
Update main_SIRG.m
