+++ modeling/SIRG/main_SIRG.m
@@ -204,6 +204,8 @@ hold off
     % axis limits
     xlim([t0 t1]);
     ylim([0 N]);
+
+    saveas(figure(1),'fig_modeling_SIRG_example1_compart.png')
     
     
 % plot NewCases (per day) of SIRG model
@@ -221,4 +223,6 @@ fig2 = scatter([1:length(NewCases)]+1,NewCases);
 
     % axis limits
     xlim([t0 length(NewCases)]+1);
+
+    saveas(figure(2),'fig_modeling_SIRG_example1_newcases.png')
 % -----------------------------------------------------------
