+++ modeling/SEIAHRD/main_SEIRAHD.m
@@ -259,6 +259,8 @@ hold off
     xlim([t0 t1]);
     ylim([0 N0]);
 
+    saveas(figure(1),'fig_modeling_SEIRAHD_example1_compart.png')
+
 
 % plot NewCases (per day) of SEIAHRD model
 figure(2)
@@ -275,4 +277,6 @@ fig2 = scatter([1:length(NewCases)]+1,NewCases);
 
     % axis limits
     xlim([t0 length(NewCases)]+1);
+
+    saveas(figure(2),'fig_modeling_SEIRAHD_example1_newcases.png')
 % -----------------------------------------------------------
