### 2022-01-20 2cca946
Update main_SEIRAHD_covid.m
