+++ modeling/SEIAHRD/main_SEIRAHD_covid.m
@@ -265,6 +265,8 @@ hold off
     xlim([t0 t1]);
     ylim([0 N0]);
 
+    saveas(figure(1),'fig_modeling_SEIRAHD_example2_compart.png')
+
 
 % plot NewCases (per day) of SEIR(+AHD) model
 figure(2)
@@ -281,4 +283,6 @@ fig2 = scatter([1:length(NewCases)]+1,NewCases);
 
     % axis limits
     xlim([t0 length(NewCases)]+1);
+
+    saveas(figure(2),'fig_modeling_SEIRAHD_example2_newcases.png')
 % -----------------------------------------------------------
