### 2022-02-03 #149 @TosinMA
Hello @dalonsoa !
About the tests:

The authors create two ways to verify the usability of the code. The first was the publication of the full code in the Code Ocean platform, which allows anyone to run all the routines e clearly see the response log and figures. It can be accessed by the url
https://codeocean.com/capsule/8871213/tree .
Moreover, there is an error script to check each function type routine (which are the RHS model functions). These are localized inside the respective subfolders. The authors are available to further clarification if necessary.
