### 2022-02-24 #149 @magsol
Hi @3mmaRand @dalonsoa:

Would you be able to provide any feedback on the fixes the authors have made? Are the fixes sufficient for publication, or are there more changes needed first?
