### 2022-03-01 #149 @dalonsoa
Hi all. Apologies for the slow reponse.

I've re-reviwed the repositorya and changes. It looks really good - definetly very dydactic; I've learnt a lot about epidemiology reading thorught the documentation! I've ticked most of the boxes above. From the end-user perspective, this is ready to go.

Now, I still think that, from a developers and sustainability perspective, this needs a bit more work:

- The checks are a great addition. I am happy to accept those as "tests" as they are now, but I would suggest to include in the documentation that those checks exist and how to run them. Even better, setup a GitHub workflow so they are automatically with every push.
- The explanation about contributing to the code is better than before, but it is standard practice to include a CONTRIBUTING file in the repository with that information instead of adding it at the end of the README.
- There is no documentation about the API (what each function does, what inputs expect and what outputs provide). This is, by far, the weakest point at the moment.

I hope this helps!
