### 2022-03-02 #149 @BrunaPavlack
Thank you so much for the feedback @dalonsoa ! We are adding suggestions. We have already created a document **contributing** and are improving it. About the API, if you can give us more suggestions on how we can do it. Thank you very much in advance!
