### 2022-03-03 #149 @dalonsoa
I'm used to Python, where documenting the API is pretty easy in several ways. It seems that for Octave, this is less straigth forward. Have a look at this comment in StackOverflow (and links therein) for a suggestion on how to do it:

https://stackoverflow.com/a/42030298/3778792

Please, give it a try and check how complex this approach is to use. It might take a while to sort that put, so I'm happy to tick this off as long as there is a clear commitment and roadmap from your side to sort that out (i.e. open an issue in the repo to tackle this).
