+++ .github/workflows/draft-pdf.yml
@@ -1,4 +1,3 @@
-name: PDFGen
 on: [push]
 
 jobs:
@@ -11,9 +10,9 @@ jobs:
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
-          journal: joss
+          journal: jose
           # This should be the path to the paper within your repo.
-          paper-path: paper.md
+          paper-path: paper/paper.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
@@ -21,4 +20,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper.pdf
+          path: paper/paper.pdf
