+++ CONTRIBUTING.md
@@ -1,3 +1,7 @@
-## Contribution to code and contact
+## Contributing
+
+The availability of open epidemiology codes contributes to the development of research and the sharing of content and tools that benefit the scientific community. Therefore, the EPIDEMIC is open to receive contributions from the community, with the aim of increasingly improving codes and teaching materials.
+
+## How to Contribute
 
 Any EPIDEMIC user can contribute the code or seek support via email americo.cunha@uerj.br or open an [issue](https://github.com/americocunhajr/EPIDEMIC/issues). You can correct spelling or grammatical errors, suggest improvements and file bugs. If you find a bug, file the issue illustrating the bug with a minimal reproducible example. Experience reports on the use of EPIDEMIC are also welcome. If you want to suggest the insertion of a compartmental model in the Modeling module, it is necessary to send a preview of the code structure following the code pattern of the other compartmental models.
