### 2022-03-09 5ac3541
Create check_error_SEIR.yml
