### 2022-03-09 3de6fc7
Create check_error_SIR.yml
