### 2022-03-09 942b629
Create check_error_SIRG.yml
