### 2022-03-09 0e587a8
Create check_error_SEIRD.yml
