+++ .github/workflows/check_error_SEIRD.yml
@@ -1,0 +1,16 @@
+name: check_error_SEIRD
+on: [push]
+jobs:
+  my-job:
+    name: check_error_SEIRD
+    runs-on: ubuntu-latest
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v2
+      - name: Set up MATLAB
+        uses: matlab-actions/setup-matlab@v1
+      - name: Run tests and generate artifacts
+        uses: matlab-actions/run-tests@v1
+        with:
+          name: check_error_SEIRD
+          path: modeling/SEIRD/check_error_SEIRD.m
