+++ CODE DOCUMENTATION.md
@@ -4,7 +4,7 @@ The EPIDEMIC is an easy-to-run educational Matlab toolkit for epidemiological an
 
 ## 🔧 Functions of module Modeling
 
-### Function rhs_SIR:
+### rhs_SIR:
 
 This function defines the system of ODEs for the SIR epidemic model.
 
@@ -38,5 +38,85 @@ Output:
 
 dydt: state rate of change - double array (4x1)
 
-### Função 02:
-- Descrição Da Função
+### rhs_SEIR:
+
+This function defines the system of ODEs for the SEIR epidemic model.
+
+The dynamic state coordinates are:
+  
+  S = susceptibles        (number of individuals)
+  
+  E = exposed             (number of individuals)
+  
+  I = infected            (number of individuals)
+ 
+  R = recovered           (number of individuals)
+  
+  C = cumulative infected (number of individuals)
+
+The epidemic model parameters are:
+
+ N  = population size   (number of individuals)
+ 
+ beta  = transmission rate (days^-1)
+ 
+ alpha = latent rate       (days^-1)
+ 
+ gamma = recovery rate     (days^-1)
+
+Inputs:
+
+ t: time                    - double
+ 
+ y: state vector            - double array (5x1)
+ 
+ param: parameters vector   - double array (4x1)
+ 
+Output:
+
+dydt: state rate of change - double array (5x1)
+
+### rhs_SEIRD:
+
+This function defines the system of ODEs for the SEIRD epidemic model.
+
+The dynamic state coordinates are:
+  
+  S = susceptibles        (number of individuals)
+  
+  E = exposed             (number of individuals)
+  
+  I = infected            (number of individuals)
+ 
+  R = recovered           (number of individuals)
+  
+  D = deceased            (number of individuals)
+  
+  C = cumulative infected (number of individuals)
+
+The epidemic model parameters are:
+
+ N0    = initial population size (number of individuals)
+ 
+ beta  = transmission rate (days^-1)
+ 
+ alpha = latent rate       (days^-1)
+ 
+ gamma = recovery rate     (days^-1)
+ 
+ delta = death rate              (days^-1)
+
+Inputs:
+
+ t: time                    - double
+ 
+ y: state vector            - double array (6x1)
+ 
+ param: parameters vector   - double array (5x1)
+ 
+Output:
+
+dydt: state rate of change - double array (6x1)
+
+
+
