### 2022-03-15 f19a9ac
Update main_SIRG.m
