### 2022-03-15 aa85edf
Update main_SIRG_example.m
