### 2022-03-15 31b0115
Update main_SIRG_example.m
