+++ CODE DOCUMENTATION.md
@@ -435,4 +435,269 @@ dydt: state rate of change - double array (8x1)
   figure 2: number of new cases in time - inplace figure
   
   
+  ### main_SEIRD_ebola
+
+ This is a main file for the SEIR epidemic model, which divides a population in 5 compartments:
+ 
+ S = susceptible
+ 
+ E = exposed
+  
+ I = infected
+ 
+ R = recovered
+ 
+ D = deceased
+
+ Infection spreads via direct contact between a susceptible and infectious individual. Delay is modeled as an exposed group: there is an latent period until an infected becomes able to transmit (infectious). Disease-related deaths are considered when infectious.
+ 
+ This model has 5 parameters:
+
+  N0    = initial population size   (number of individuals)
+  
+  beta  = transmission rate         (days^-1)
+  
+  alpha = latent rate               (days^-1)
+  
+  gamma = recovery rate             (days^-1)
+  
+  delta = death rate                (days^-1)
+  
+ This codes uses rhs_SEIRD.m to define the ODE system and outputs the plots and R_nought value. Calculations are made on a day time scale. The parameters used in this example were taken from this work: P. Diaz, P. Constantine, K. Kalmbach, E. Jones and S. Pankavichm. A modified SEIR model for the spread of Ebolain Western Africa and metrics for resource allocation, Applied Mathematics and Computation, 324(1): 141-155, 2018
+ 
+ Inputs:
+  
+  param: parameters vector      - double array (5x1)
+  
+  IC: initial conditions vector - double array (6X1)
+  
+  tspan: time interval          - double array (?x1)
+  
+  rhs_SEIRD: SEIRD equations file   - .m function file
+
+  Outputs:
+  
+  R_nought: basic reproduction number   - double
+  
+  figure 1: model state in time         - inplace figure
+  
+  figure 2: number of new cases in time - inplace figure
+  
+  
+ ### main_SIRG
+
+ This is the main file for a economic-epidemic coupled dynamic model, which divides a population in 4 compartments:
+
+ S = susceptible
+  
+ I = infected
+ 
+ R = recovered
+ 
+ G = guarded
+
+ Infection spreads via direct contact between a susceptible and infected individual, with no delay. No deaths are considered, all infected become recovered.
+ 
+ This model has 7 parameters:
+
+   N     = population size   (number of individuals)
+   
+   beta  = transmission rate (days^-1)
+   
+   gamma = recovery rate     (days^-1)
+   
+   phi   = quantine-in rate  (days^-1)
+   
+   theta = quantine-out rate (days^-1)
+   
+   tq0   = quantine start (days)
+   
+   tq1   = quantine end (days)
+  
+  This codes uses rhs_SIRG.m to define the ODE system and outputs the plots and R_nought value. Calculations are made on a day time scale.
+ 
+ Inputs:
+  
+  param: parameters vector      - double array (4x1)
+  
+  IC: initial conditions vector - double array (5X1)
+  
+  tspan: time interval          - double array (?x1)
+  
+  rhs_SIRG: SIRG equations file   - .m function file
+
+  Outputs:
+  
+  R_nought: basic reproduction number   - double
+  
+  figure 1: model state in time         - inplace figure
+  
+  figure 2: number of new cases in time - inplace figure
+
+
+ ### main_SIRG_example
+
+ This is the main file for a economic-epidemic coupled dynamic model, which divides a population in 4 compartments:
+
+ S = susceptible
+  
+ I = infected
+ 
+ R = recovered
+ 
+ G = guarded
+
+ Infection spreads via direct contact between a susceptible and infected individual, with no delay. No deaths are considered, all infected become recovered.
+ 
+ This model has 7 parameters:
+
+   N     = population size   (number of individuals)
+   
[…]
