### 2022-03-17 #149 @BrunaPavlack
@dalonsoa and @3mmaRand thank you so much for all the contributions! We added a document about the contribution and as the code is used we improve. It was also added to the tutorials about test routines. Regarding the code documentation, we made a file called CODE DOCUMENTATION.md where we explained the functions used in the compartmental models and the inputs and outputs of the codes of the modules Modeling, Trends and Forecasts. We are currently updating this document in line with code improvements. If the user needs information about the functions in Octave, it is possible to put, for example, help rhs_SIR and information about the function will appear on the screen. This information will also be added in the tutorials.
