### 2022-04-05 #149 @labarba
I see the recommendation to accept from  @3mmaRand, in a [comment above](https://github.com/openjournals/jose-reviews/issues/149#issuecomment-1059941543), but some items of the [checklist](https://github.com/openjournals/jose-reviews/issues/149#issue-1051276200) remain unticked. Did you forget to finish ticking those, @3mmaRand, and could you do so now, just to leave the review thread more tidy?
