### 2022-04-15 #149 @labarba
On the paper, did you really mean to use the word "retrieved" on this line? It doesn't make complete sense to me.
https://github.com/americocunhajr/EPIDEMIC/blob/45168e8a8249a0a04f4a818c9b9cbcdd69ecc4dc/paper/paper.md?plain=1#L134

Also, the whole parenthetical here:
[(eg infected vs time / infected by 1M inhab vs time, when normalizing the curves approach the same level, without local ](https://github.com/americocunhajr/EPIDEMIC/blob/45168e8a8249a0a04f4a818c9b9cbcdd69ecc4dc/paper/paper.md?plain=1#L150)
…reads like unfinished text or draft notes. Please fix.

On line [153](https://github.com/americocunhajr/EPIDEMIC/blob/45168e8a8249a0a04f4a818c9b9cbcdd69ecc4dc/paper/paper.md?plain=1#L153), "Bellow" should be "Below" — also just say "Below is a brief description" (delete "it" and "presented").

Wherever you used "the EPIDEMIC" you should probably change to "the EPIDEMIC software" or "the EPIDEMIC code" or just "EPIDEMIC" (i.e., the article is not needed is you are only using the name).

Line [173](https://github.com/americocunhajr/EPIDEMIC/blob/45168e8a8249a0a04f4a818c9b9cbcdd69ecc4dc/paper/paper.md?plain=1#L173): "It is interesting that the user has previous knowledge" >> this is badly formed. Do you mean to say "It is necessary"? Or something else? Maybe just say "The user should have previous knowledge…"

Line [176](https://github.com/americocunhajr/EPIDEMIC/blob/45168e8a8249a0a04f4a818c9b9cbcdd69ecc4dc/paper/paper.md?plain=1#L176): "The figure 1: >> "Figure 1…" (no article).

Line [192](https://github.com/americocunhajr/EPIDEMIC/blob/45168e8a8249a0a04f4a818c9b9cbcdd69ecc4dc/paper/paper.md?plain=1#L192): "analyzes" >> "analyses" (for the noun form)

L193: "In Figure [2] is possible to see" >> "Figure [2] shows"

L206: "The Figure [3] shows" >> "Figure [3] shows"

L207: "This graph shows the total number of deaths per 1M inhabitants evolution of the COVID-19 pandemic in several countries." >> This graph shows the evolution of the total number of deaths per 1M inhabitants in several countries.
(Also fix the figure caption.)

L225: "The Figure [4] shows" >> "Figure [4] shows" and also "at EPIDEMIC" should be "by EPIDEMIC" (fix caption too).
