### 2022-04-15 #149 @BrunaPavlack
> Hi @BrunaPavlack 👋 — I see that the release tag is `v.1.0.0-beta`, whereas here we have marked it as `v.1.0.0`. These two should match. Which one is it? (The release is also from January. Was there no change since?)

Hi!!! I created a new tag with the updated version.
