+++ paper/paper.md
@@ -98,9 +98,9 @@ Observatório Fluminense* (COVID-19RJ) [@covid19-rj], which aims to
 monitor the pandemic progress in Brazil, make reliable predictions about
 the short term evolution and provide high-quality educational material
 about the mathematical modeling and analysis of COVID-19. To support the
-research carried out in the COVID-19RJ project, the EPIDEMIC was
+research carried out in the COVID-19RJ project, EPIDEMIC was
 developed, because through this code it is possible to generate
-monitoring graphics, trends and disease forecasts. The EPIDEMIC is cited
+monitoring graphics, trends and disease forecasts. EPIDEMIC is cited
 in the reports of the COVID-19RJ project.
 
 In the context of the COVID-19 pandemic, it was observed that
@@ -120,25 +120,25 @@ In a mathematical biology course, for example, students are exposed to the deriv
 from various compartmental models based on nonlinear differential equations, so that the vast majority 
 have no analytical solution. In this context, there is a natural need to use numerical methods to obtain 
 the answer from the mathematical model. Packages like GNU Octave and Matlab have very robust ordinary differential equations (ODEs) solvers that 
-make this task much easier. The EPIDEMIC tutorials provide pedagogical activities in this sense, as they illustrate,
+make this task much easier. EPIDEMIC tutorials provide pedagogical activities in this sense, as they illustrate,
 via relatively streamlined and very well organized codes, how to perform the numerical integration of these models based on ODEs. 
 The programming structure followed in the tutorial, with several compartmental models, also aims to show students how easy it 
 is to carry out such numerical integration for another type of compartmental model that the student can find in the literature 
 or develop himself/herself.
 
-The benefits of the EPIDEMIC tutorial go beyond the simple exercise of programming numerical integration, 
+The benefits of EPIDEMIC tutorial go beyond the simple exercise of programming numerical integration, 
 the analysis of the response curves of different models allows students to develop intuition about the dynamic 
 behavior of dynamic epidemiological systems. For example, when seeing in the same figure the 3 curves of the SIR 
 model (susceptible, infected and recovered), the student can see that the initial growth of an outbreak is
 accompanied at the same time by a significant reduction in the number of susceptibles, being succeeded by an 
-increase in the number of retrieved. An interesting analogy in this case that aids in the students' 
+increase in the number of recovereds. An interesting analogy in this case that aids in the students' 
 understanding is the following: be the infected as the intensity of a fire, the susceptible as the available firewood, 
 and the recovered as the burnt firewood after a fire. At the beginning there is a lot of firewood available, 
 a large fire ensues, then, as the firewood burns, the intensity of the fire decreases.
 
-The EPIDEMIC codes have already been used in a course on numerical and computational methods in Rio de Janeiro State University to present the 
+EPIDEMIC codes have already been used in a course on numerical and computational methods in Rio de Janeiro State University to present the 
 part of polynomial regression and exponential curve fitting. In this course, the code was used on three fronts, 
-according to the EPIDEMIC modules: modeling, trends and forecast. In the module modeling, the compartmental models were 
+according to EPIDEMIC modules: modeling, trends and forecast. In the module modeling, the compartmental models were 
 presented as computer simulators, students run the codes as a black box to familiarize themselves with simulation basics,
 and get acquainted with the Octave software. The code is also used in interpolation classes. We want to obtain a 
 polynomial function that describes the number of infected people at the peak of the epidemic as a function of the transmission 
@@ -147,34 +147,29 @@ of infected at peak. Then they interpolate polynomial curves to describe this fu
[…]
