### 2022-04-22 b1c17f0
slight copy editing
For JOSE publication https://github.com/openjournals/jose-reviews/issues/149
