### 2022-04-22 #149 @labarba
Please merge my PR with some slight copy editing: https://github.com/americocunhajr/EPIDEMIC/pull/4
