### 2022-04-26 #149 @labarba
I don't understand. I did copy editing on the paper, and submitted a PR, but it was closed without merging: https://github.com/americocunhajr/EPIDEMIC/pull/4
