### 2022-04-27 #149 @labarba
@BrunaPavlack – Could you please check the proof? I think there's a figure missing. See the string
`![Trend graph of total number of deaths per 1M inhabitants in several countries generated in EPIDEMIC.]` at the end of the **Trends** section.
