### 2022-04-27 #149 @labarba
The only issue that remain is a URL that bleeds into the margin in the References. Not sure right now how that can be fixed. In pure LaTeX, I would use the `\url` tag (which will break long URLs as needed into the next line). If you want to try a few things, you can preview your compiled paper using the service at: http://whedon.theoj.org
