+++ paper/paper.bib
@@ -20,7 +20,7 @@
     site = {Towards Data Science},
     title = {Transparency, Reproducibility, and Validity of {COVID-19} Projection Models},
     year = {2020},
-    url = {https://towardsdatascience.com/transparency-reproducibility-and-validity-of-covid-19-projection-models-78592e029f28},
+    url = {encurtador.com.br/jGOR4},
     urldate = {2020-05-04}
 }
 
